refactor sort_table_for_interval_join
